Fixed Money bug with comma's Fixed Query not seperating users

+++ iSpendApp/iSpendDAL/Dto/BillDto.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.SqlTypes;
 using System.Text;
 using iSpendInterfaces;
 
@@ -7,7 +8,7 @@ namespace iSpendDAL.Dto
 {
     internal class BillDto:IBill
     {
-        public BillDto(int billId, string billName, decimal balance,DateTimeOffset time)
+        public BillDto(int billId, string billName, double balance,DateTimeOffset time)
         {
             BillId = billId;
             BillName = billName;
@@ -17,10 +18,11 @@ namespace iSpendDAL.Dto
 
         public int BillId { get; set; }
         public string BillName { get; set; }
-        public decimal BillBalance { get; set; }
+        public double BillBalance { get; set; }
         public IEnumerable<ITransaction> Transactions { get; set; }
         public int IconId { get; set; }
         public IEnumerable<int> AccountIds { get; set; }
         public DateTimeOffset DateOfCreation { get; set; }
+
     }
 }
